+++ ChitChatClient/Windows/ChatWindow.xaml.cs
@@ -38,92 +38,95 @@ namespace ChitChatClient.Windows
 			// Initializing user refresh timer
 			userRefreshTimer = new DispatcherTimer();
 
-			userRefreshTimer.Tick += new EventHandler(RefreshServerUsersTimed);
+			userRefreshTimer.Tick += new EventHandler(RefreshServerUsersTask);
 			userRefreshTimer.Interval = new TimeSpan(0, 0, 5);
 			userRefreshTimer.Start();
 		}
 
-		private int SelectedServerId
+		private Server? SelectedServer
 		{
 			get
 			{
 				Server selectedServer = (Server)serverListView.SelectedItem;
 				if (selectedServer == null)
 				{
-					return -1;
+					return null;
 				}
 				else
 				{
-					return selectedServer.Id;
+					return selectedServer;
 				}
 			}
 		}
 
-		private async Task RefreshJoinedServers()
+		private void RefreshSeverListView()
 		{
-            // Get all joined servers
-            await chatInterface.GetServers();
-            serverListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
-            serverListView.SelectedIndex = 0;
-			if (chatInterface.Servers.Count == 0)
-			{
-                removeServerButton.IsEnabled = false;
-            }
-            else
-			{
-                removeServerButton.IsEnabled = true;
-            }
-        }
+			serverListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
+			CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Filter = FilterServer;
+			CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Refresh();
+			serverListView.SelectedIndex = 0;
 
-		private async Task RefreshServerUsers()
-		{
-			await chatInterface.GetServerUsers(SelectedServerId);
-			userListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
-
-			Server selectedServer = (Server)serverListView.SelectedItem;
-			int onlineUserCount = await serverService.GetOnlineUserCount(selectedServer.Id);
-
-			onlineUserCountTB.Text = "Online " + onlineUserCount.ToString();
-			totalUserCountTB.Text = "Total " + selectedServer.UserCount.ToString();
+			removeServerButton.IsEnabled = chatInterface.Servers.Count != 0;
 		}
 
-		private async void RefreshServerUsersTimed(object? sender, EventArgs e)
+		private void RefreshUserListView(int? onlineUserCount, int? totalUserCount)
 		{
-			await RefreshServerUsers();
+			userFilterTextBox.Text = string.Empty;
+			userListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
+			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Filter = FilterUser;
+			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Refresh();
+			userListView.SelectedIndex = -1;
+
+			if (onlineUserCount != null && totalUserCount != null)
+			{
+				onlineUserCountTB.Text = "Online " + onlineUserCount.ToString();
+				totalUserCountTB.Text = "Total " + totalUserCount.ToString();
+			}
+		}
+
+		private async Task RefreshServerUsersWrapper()
+		{
+			await chatInterface.RefreshServerUsers(RefreshUserListView, SelectedServer);
+		}
+
+		private async void RefreshServerUsersTask(object? sender, EventArgs e)
+		{
+			await RefreshServerUsersWrapper();
 		}
 
         private async void Window_Loaded(object sender, RoutedEventArgs e)
         {
-            await RefreshJoinedServers();
-            await RefreshServerUsers();
-        }
+            await chatInterface.RefreshJoinedServers(RefreshSeverListView);
+            await chatInterface.RefreshServerUsers(RefreshUserListView, SelectedServer);
+			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Filter = FilterUser;
+			CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Filter = FilterServer;
+		}
 
-        private async void createServerButton_Click(object sender, RoutedEventArgs e)
+		private async void createServerButton_Click(object sender, RoutedEventArgs e)
 		{
 			// Create a server
 			CreateServerWindow createServerWindow = new();
 			createServerWindow.ShowDialog();
-			await RefreshJoinedServers();
-        }
+			await chatInterface.RefreshJoinedServers(RefreshSeverListView);
+		}
 
 		private async void removeServerButton_Click(object sender, RoutedEventArgs e)
 		{
-			// Remove a server
+			// Leave a server
 			if (serverListView.SelectedIndex != -1)
 			{
-				bool result = await chatInterface.LeaveServer(SelectedServerId);
+				bool result = await chatInterface.LeaveServer(SelectedServer);
 
 				if (result)
 				{
 					MessageBox.Show("Server left successfully", "Success", MessageBoxButton.OK, MessageBoxImage.Information);
-                    await RefreshJoinedServers();
 				}
 				else
 				{
 					MessageBox.Show("Failed to leave server", "Error", MessageBoxButton.OK, MessageBoxImage.Error);
 				}
 
-				await RefreshJoinedServers();
+				await chatInterface.RefreshJoinedServers(RefreshSeverListView);
                 serverListView.SelectedIndex = 0;
 			}
         }
@@ -133,7 +136,7 @@ namespace ChitChatClient.Windows
 		{
 			ServerListWindow serverListWindow = new();
 			serverListWindow.ShowDialog();
-			await RefreshJoinedServers();
+			await chatInterface.RefreshJoinedServers(RefreshSeverListView);
         }
 
 		private void serverInfoButton_Click(object sender, RoutedEventArgs e)
@@ -164,13 +167,58 @@ namespace ChitChatClient.Windows
 
 		private async void serverListView_SelectionChanged(object sender, SelectionChangedEventArgs e)
 		{
-			await chatInterface.GetServerUsers(SelectedServerId);
-            userListView.GetBindingExpression(ListView.ItemsSourceProperty).UpdateTarget();
+			await chatInterface.RefreshServerUsers(RefreshUserListView, SelectedServer);
         }
 
 		private void Window_Closing(object sender, System.ComponentModel.CancelEventArgs e)
 		{
 			userRefreshTimer?.Stop();
+		}
+
+		private void userFilterTextBox_TextChanged(object sender, TextChangedEventArgs e)
+		{
+			CollectionViewSource.GetDefaultView(userListView.ItemsSource).Refresh();
+		}
+
+		private void filterServerTextBox_TextChanged(object sender, TextChangedEventArgs e)
+		{
+			CollectionViewSource.GetDefaultView(serverListView.ItemsSource).Refresh();
+		}
+
+		private bool FilterUser(object obj)
+		{
+			if (string.IsNullOrEmpty(userFilterTextBox.Text))
+			{
+				return true;
+			}
+
+			if (obj is User user)
+			{
+				if (user.Username.ToLower().Contains(userFilterTextBox.Text.ToLower()))
+				{
+					return true;
+				}
+			}
+
+			return false;
+		}
+
+		private bool FilterServer(object obj)
+		{
+			if (string.IsNullOrEmpty(filterServerTextBox.Text))
+			{
+				return true;
+			}
+
+			if (obj is Server server)
+			{
+				if (server.Name.ToLower().Contains(filterServerTextBox.Text.ToLower()))
+				{
+					return true;
+				}
+			}
+
+			return false;
 		}
 	}
 }