+++ release_notes.md
@@ -2,10 +2,17 @@
 
 - Check for new version when starting the client and notify the user if a new version is available
 - Added on status to user list which refreshes every 5 seconds
+- Added online and total user count to server join list
+- Added ability to sort servers by join status, name and user count
+- Added online and total user count to user list
+- Added ability to filter servers by name on chat screen
+- Added ability to filter users by name on chat screen
+- Added automatic sorting for users by their online status
 
 ## Changes
 
 - Increased chat window size to 1000x540
+- Changed the way server list is displayed when joining servers
 
 ## Fixes